model review

+++ PetCareFinalVersion/PetCareFinalVersion/Data/AbstractAnimal.cs
@@ -19,7 +19,7 @@ namespace PetCareFinalVersion.Data
         public abstract string Type { get; set; }
         public abstract string Breed { get; set; }
         public abstract string Age { get; set; }
-        public abstract int Weight { get; set; }
+        public abstract float Weight { get; set; }
         public abstract string Status { get; set; }
         public abstract string Size { get; set; }
         public abstract string Description { get; set; }
